requirements.txt
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
smplx==0.1.28
trimesh==4.4.8
pyrender==0.1.45
mediapipe==0.10.14
opencv-python==4.10.0.84
xgboost==2.1.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pyarrow==17.0.0
joblib==1.4.2
